取消道具统计的定时任务

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/player/service/impl/GamePlayerItemLogServiceImpl.java
@@ -41,7 +41,7 @@ public class GamePlayerItemLogServiceImpl extends ServiceImpl<GamePlayerItemLogM
     private ConfigDataService configDataService;
 
     @Override
-    public GamePlayerItemLog writePlayerItemLog(long serverId, BackpackLog backpacklog) {
+    public GamePlayerItemLog writePlayerItemLog(Integer serverId, BackpackLog backpacklog) {
         return new GamePlayerItemLog().setServerId(serverId).setPlayerId(backpacklog.getPlayerId()).setType(backpacklog.getType())
                 .setWay(backpacklog.getWay()).setSyncTime(backpacklog.getCreateDate()).setItemId(backpacklog.getItemId())
                 .setBeforeNum(backpacklog.getBeforeNum()).setAfterNum(backpacklog.getAfterNum())
